+++ core/config/config.go
@@ -83,8 +83,6 @@ type Configuration struct {
 	Servers               ServerConfigurations
 	OperationServers      map[string]ServerConfigurations
 	HTTPClient            *http.Client
-	// Deprecated: retry options were removed to reduce complexity of the client. If this functionality is needed, you can provide your own custom HTTP client. An option to add HTTP middleware will be introduced soon. This field has no effect, and will be removed in a later update
-	RetryOptions *clients.RetryConfig //nolint:staticcheck //will be removed in a later update
 
 	// If != nil, a goroutine will be launched that will refresh the service account's access token when it's close to being expired.
 	// The goroutine is killed whenever this context is canceled.
@@ -94,6 +92,8 @@ type Configuration struct {
 
 	// Deprecated: validation using JWKS was removed, for being redundant with token validation done in the APIs. This field has no effect, and will be removed in a later update
 	JWKSCustomUrl string `json:"jwksCustomUrl,omitempty"`
+	// Deprecated: retry options were removed to reduce complexity of the client. If this functionality is needed, you can provide your own custom HTTP client. This field has no effect, and will be removed in a later update
+	RetryOptions *clients.RetryConfig //nolint:staticcheck //will be removed in a later update
 
 	setCustomEndpoint bool
 }
@@ -232,21 +232,21 @@ func WithToken(token string) ConfigurationOption {
 	}
 }
 
-// Deprecated: retry options were removed to reduce complexity of the client. If this functionality is needed, you can provide your own custom HTTP client. An option to add HTTP middleware will be introduced soon. This option has no effect, and will be removed in a later update
+// Deprecated: retry options were removed to reduce complexity of the client. If this functionality is needed, you can provide your own custom HTTP client. This option has no effect, and will be removed in a later update
 func WithMaxRetries(_ int) ConfigurationOption {
 	return func(config *Configuration) error {
 		return nil
 	}
 }
 
-// Deprecated: retry options were removed to reduce complexity of the client. If this functionality is needed, you can provide your own custom HTTP client. An option to add HTTP middleware will be introduced soon. This option has no effect, and will be removed in a later update
+// Deprecated: retry options were removed to reduce complexity of the client. If this functionality is needed, you can provide your own custom HTTP client. This option has no effect, and will be removed in a later update
 func WithWaitBetweenCalls(_ time.Duration) ConfigurationOption {
 	return func(config *Configuration) error {
 		return nil
 	}
 }
 
-// Deprecated: retry options were removed to reduce complexity of the client. If this functionality is needed, you can provide your own custom HTTP client. An option to add HTTP middleware will be introduced soon. This option has no effect, and will be removed in a later update
+// Deprecated: retry options were removed to reduce complexity of the client. If this functionality is needed, you can provide your own custom HTTP client. This option has no effect, and will be removed in a later update
 func WithRetryTimeout(_ time.Duration) ConfigurationOption {
 	return func(config *Configuration) error {
 		return nil